[manager] Deal with API changes
The new ResourceValueMap overrides some of the Map functions. However,
it doesn't override the putAll method which is problematic since it
overrides the keySet and returns its own local instance of keys which
is only updated during the put method. Using putAll results in KeySet
return 0 regardless of how many elements are in the map.